SparkDelta/requirements.txt
pandas==2.2.2
numpy==2.0.2
pyspark==3.5.2
delta-spark==3.2.0
kafka-python==2.0.2